Change file structuring
